requirements.txt
streamlit
requests
json
datetime
matplotlib.pyplot 
pandas 
matplotlib.dates 